+++ CHANGELOG.md
@@ -1,4 +1,6 @@
-##### Fri Nov 03 2017 Sundeep Anand <suanand@redhat.com> 
+##### Tue Nov 07 2017 Sundeep Anand <suanand@redhat.com> 
+* Modify APIs - `status` is `package` and `workload` is `release` now
+* Add `releases` and `packages` reports to landing page
 * Improve UI, Jobs tab, package config with stats, better navigation
 
 ##### Fri Oct 27 2017 Sundeep Anand <suanand@redhat.com> - release_0.1.4